menu

+++ src/menu.cpp
@@ -24,14 +24,13 @@ bool isReachedTarget = false;
 bool timeForText = false;
 
 // Text messages
-const char* messages[] = {
+const char *messages[] = {
     "Hello There!!! Ready to conquer?",
     "I hope you are ready. But hey don't hit (START) just yet!",
     "Master the game by first mastering the rules.",
     "Take a moment to check the about page - it could save you from a few extra lives",
     "Good luck with the Game.",
-    " "
-};
+    " "};
 
 Font font = GetFontDefault();
 int messageIndex = 0;
@@ -54,13 +53,12 @@ void InitMenu()
 
     // Initialize positions based on loaded textures
     initialPosition = (Vector2){0, 500};
-  // Define the right offset
-float offsetRight = 200.0f;  // Adjust this value as needed
+    // Define the right offset
+    float offsetRight = 200.0f; // Adjust this value as needed
 
-// Update positions with the offset
-targetPosition = (Vector2){1280 / 2.0f - broom.width / 2.0f + offsetRight, initialPosition.y};
-standPosition = (Vector2){1280 / 2.0f - stand.width / 2.0f + offsetRight, 700 - stand.height};
-
+    // Update positions with the offset
+    targetPosition = (Vector2){1280 / 2.0f - broom.width / 2.0f + offsetRight, initialPosition.y};
+    standPosition = (Vector2){1280 / 2.0f - stand.width / 2.0f + offsetRight, 700 - static_cast<float>(stand.height)};
 }
 
 void UpdateMenu()
@@ -68,30 +66,36 @@ void UpdateMenu()
     UpdateMusicStream(musicm);
 
     // Broom movement logic
-    if (isBroom && initialPosition.x < targetPosition.x) {
+    if (isBroom && initialPosition.x < targetPosition.x)
+    {
         initialPosition.x += 200.0f * GetFrameTime();
     }
 
-    if (initialPosition.x >= targetPosition.x && !isReachedTarget) {
+    if (initialPosition.x >= targetPosition.x && !isReachedTarget)
+    {
         isReachedTarget = true;
         timeElapsed += GetFrameTime();
 
-        if (timeElapsed >= delayTime) {
+        if (timeElapsed >= delayTime)
+        {
             isBroom = false;
             timeForText = true;
         }
     }
 
     // Typing effect logic for messages
-    if (!isBroom && timeForText) {
-        if (IsKeyPressed(KEY_SPACE) && messageIndex < totalMessages - 1) {
+    if (!isBroom && timeForText)
+    {
+        if (IsKeyPressed(KEY_SPACE) && messageIndex < totalMessages - 1)
+        {
             messageIndex++;
             charIndex = 0;
             timePassed = 0.0f;
         }
 
         timePassed += GetFrameTime();
-        if (timePassed >= typingSpeed && charIndex < strlen(messages[messageIndex])) {
+        if (timePassed >= typingSpeed && charIndex < static_cast<int>(strlen(messages[messageIndex])))
+        {
             charIndex++;
             timePassed = 0.0f;
         }
@@ -104,22 +108,28 @@ void DrawMenu()
     DrawTexture(backgroundm, 0, 0, WHITE);
 
     // Draw broom or stand based on state
-    if (isBroom) {
+    if (isBroom)
+    {
         DrawTexture(broom, initialPosition.x, initialPosition.y, WHITE);
-    } else {
+    }
+    else
+    {
         DrawTexture(stand, standPosition.x, standPosition.y, WHITE);
     }
 
     // Draw text with word wrapping and typing effect
     int currentCharIndex = 0;
     int line = 0;
-    for (int i = 0; i < charIndex && line < MAX_LINES; i++) {
-        if (currentCharIndex >= MAX_CHARACTERS_PER_LINE || messages[messageIndex][i] == '\n') {
+    for (int i = 0; i < charIndex && line < MAX_LINES; i++)
+    {
+        if (currentCharIndex >= MAX_CHARACTERS_PER_LINE || messages[messageIndex][i] == '\n')
+        {
             line++;
             currentCharIndex = 0;
         }
 
-        if (line < MAX_LINES) {
+        if (line < MAX_LINES)
+        {
             DrawTextEx(font, TextFormat("%c", messages[messageIndex][i]),
                        (Vector2){textArea.x + currentCharIndex * 18, textArea.y + line * lineHeight}, 30, 2, DARKGRAY);
             currentCharIndex++;
@@ -127,7 +137,8 @@ void DrawMenu()
     }
 
     // Instruction to proceed to the next message
-    if (timeForText && messageIndex < totalMessages - 1 && charIndex == strlen(messages[messageIndex])) {
+    if (timeForText && messageIndex < totalMessages - 1 && charIndex == static_cast<int>(strlen(messages[messageIndex])))
+    {
         DrawText("Press SPACE for next message", 900, 650, 20, LIGHTGRAY);
     }
 }
